Extract some common Gradle task management functionality into the `common` module.
